AESD.Records.Server

Description

A skeletal implementation of a server.

Entry ponts

serverMain

Arguments

:: ModelManager a

=> String
The WebSocket host address.

-> Int
The WebSocket port number.

-> Maybe Int
The number of records per chunk.

-> a
The inital state of the model manager.

-> IO ()
Action to run the server.

Run the server.

class ModelManager a where

Type class for managing a model.

Methods

listModels :: ServiceM a [ModelMeta]

List all models.

lookupModel

Arguments

:: ModelIdentifier
The model identifier.

-> ServiceM a ModelMeta
Action to lookup a model.

Lookup a model.

loadContent

Arguments

:: ModelMeta
The model metadata.

-> Maybe (Either BookmarkIdentifier Filter)
Maybe the bookmark or filter.

-> [VariableIdentifier]
The variables to select.

-> Maybe Int
The maximum number of records to return.

-> Maybe Int
The number of records per chunk.

-> ServiceM a (ServiceM a [RecordContent])
Action returning an action for loading chunks of record data.

Load record data.

listBookmarks

Arguments

:: ModelIdentifier
The model identifier.

-> ServiceM a [BookmarkMeta]
Action to list all bookmarks.

List all bookmarks.

lookupBookmark

Arguments

:: ModelIdentifier
The model identifier.

-> BookmarkIdentifier
The bookmark identifier.

-> ServiceM a BookmarkMeta
Action to lookup a bookmark.

Lookup a bookmark.

saveBookmark

Arguments

:: ModelIdentifier
The model identifier.

-> BookmarkMeta
The bookmark metadata.

-> ServiceM a BookmarkMeta
Action to save a bookmark and return it.

Save a bookmark.

doWork

Arguments

:: ModelMeta
The model metadata.

-> [VarValue]
The values of the input variables.

-> Maybe Int
The maximum number of records to return.

-> Maybe Int
The number of records per chunk.

-> ServiceM a (ServiceM a [RecordContent])
Action to perform the work and return an action for loading chunks of the new records.

Perform work for a model.

Service monad

data ServiceM s a

The service.

serviceM :: MonadTrans t => ServiceM s a -> t (ServiceM s) a

Lift into the service.

fromService

Arguments

:: (s -> a)
The function.

-> ServiceM s a
The action resulting from applying the function to the state.

Apply a function to a service.

runServiceToIO

Arguments

:: TVar s
The initial state.

-> ServiceM s a
The service.

-> IO (Either String a)
The action to run the service, possibly reporting an error.

Run the service to the IO monad

Modifying a service

modifyService :: (s -> s) -> ServiceM s ()

Modify the state in a service.

modifyService' :: (s -> Either String (s, a)) -> ServiceM s a

Modify the state in a service, allowing for an error result.

modifyServiceIO' :: (s -> IO (Either String (s, a))) -> ServiceM s a

Modify the state in a service, allowing for an error result.