AESD.Records.Server.File

Description

Support for file-based serving of records.

Metadata

buildVarMetas

Arguments

:: [[String]]
The rows of data.

-> [VarMeta]
The variable metadata for each column of the table.

Construct variable metadata for a table of data. The strings in the table are analyzed to determine whether each column is a real, integer, or string value. If the name of the variable contains its units in square brackets, the units are separated from the variable and put in the units metadata.

buildVarMeta

Arguments

:: VariableIdentifier
The variable identifier.

-> String
The name of the variable.

-> VarType
The type for the variable.

-> VarMeta
The variable metadata.

Construct Variable metadata. If the name of the variable contains its units in square brackets, the units are separated from the variable and put in the units metadata.

Data

buildModelContent

Arguments

:: [[String]]
The rows of data.

-> [RecordContent]
The records.

Construct records from a table of data. The strings in the table are parsed into values with the most specific value type in each column.