AESD.Types.Bookmark

Description

Types for bookmarks, which are named sets of records or a filter for records.

Use makeInterval, makeSet, or makeFilter to construct a bookmark, and identifier, name, and onBookmarkMeta to access its contents. The function filterWithBookmark will filter a set of records, yielding those that satisfy the bookmark.

Bookmarks

type BookmarkIdentifier = String

Unique identifier for a bookmark.

data BookmarkMeta

Metadata for a bookmark.

makeInterval

Arguments

:: Maybe BookmarkIdentifier
The unique identifier for the bookmark, optional for a new bookmark.

-> String
The name for the bookmark.

-> (Maybe RecordIdentifier, Maybe RecordIdentifier)
The interval for the record identifiers in the bookmark. The interval may be open on either side of the interval.

-> BookmarkMeta
The bookmark.

Construct a bookmark for an interval of record identifiers.

makeSet

Arguments

:: Maybe BookmarkIdentifier
The unique identifier for the bookmark, optional for a new bookmark.

-> String
The name for the bookmark.

-> [RecordIdentifier]
The set of record identifiers in the bookmark.

-> BookmarkMeta
The bookmark.

Construct a bookmark for a set of record identifiers.

makeFilter

Arguments

:: Maybe BookmarkIdentifier
The unique identifier for the bookmark, optional for a new bookmark.

-> String
The name for the bookmark.

-> Filter
The expression for filtering records for the bookmark.

-> BookmarkMeta
The bookmark.

Construct a bookmark for a filter.

setIdentifier :: BookmarkIdentifier -> BookmarkMeta -> BookmarkMeta

Set the unique identifier of a bookmark. This is useful for adding the identifier to a new bookmark supplied by a client, but lacking a unique identifier.

identifier :: BookmarkMeta -> Maybe BookmarkIdentifier

Get the unique identifier of a bookmark.

name :: BookmarkMeta -> String

Get the name of a bookmark.

onBookmarkMeta

Arguments

:: ((Maybe RecordIdentifier, Maybe RecordIdentifier) -> a)
Handle an interval of record identifiers.

-> ([RecordIdentifier] -> a)
Handle a set of record identifiers.

-> (Filter -> a)
Handle a filter.

-> a
Default result.

-> BookmarkMeta
The bookmark.

-> a
The result of applying the first applicable handler to the bookmark.

Apply functions to a bookmark. A valid bookmark will only specify an interval, set, or filter, so only one of the handler functions will be applied.

For example:

>>> import AESD.Types.Bookmark (makeSet, onBookmarkMeta)
>>> 
>>> let b = makeSet (Just "42") "sample set bookmark" [1066, 1812, 1939]
>>> onBookmarkMeta (const "interval") (const "set") (const "filter") "unknown" b
"set"
>>> onBookmarkMeta (const []) id (const []) [] b
[1066,1812,1939]

filterWithBookmark

Arguments

:: BookmarkMeta
The bookmark.

-> [RecordContent]
The original records.

-> [RecordContent]
The records satisfying the bookmark.

Find the records satisfying a bookmark.

Collections of bookmarks

data BookmarkMetas

A collections of bookmarks.

makeBookmarks :: [BookmarkMeta] -> BookmarkMetas

Construct a collection of bookmarks.

bookmarks :: BookmarkMetas -> [BookmarkMeta]

Get the bookmarks in a collection.