AESD.Types.Domain

Description

Types for domains of variables.

Use makeInterval or makeSet to construct a domain, and variable or onDomainMeta to access its contents. Records can be filtered by domain using filterDomain and filterDomains

Variable domains

data DomainMeta

Metadata for the domain of a variable.

makeInterval

Arguments

:: VariableIdentifier
The identifier for the variable in question.

-> (Maybe DataValue, Maybe DataValue)
The interval of values in the domain. The interval may be open on either side of the interval.

-> DomainMeta
The variable's domain.

Construct a domain for an interval of values.

makeSet

Arguments

:: VariableIdentifier
The identifier for the variable in question.

-> [DataValue]
The set of values in the domain.

-> DomainMeta
The variable's domain.

Construct a domain for a set of values.

variable :: DomainMeta -> VariableIdentifier

Get the variable identifier for a domain.

onDomainMeta

Arguments

:: (VariableIdentifier -> (Maybe DataValue, Maybe DataValue) -> a)
Handle an interval of values.

-> (VariableIdentifier -> [DataValue] -> a)
Handle a set of values.

-> a
The default result.

-> DomainMeta
The variable's domain.

-> a
The result of applying the first applicable handler to the domain.

Apply a function to a variable's domain. A valid domain will only specify an interval or a set, so only one of the handler function will be applied.

For example:

>>> import AESD.Types.Domain (onDomainMeta)
>>> import AESD.Types.Value (realValue)
>>> 
>>> let d = makeSet 42 $ realValue <$> [1066, 1812, 1939]
>>> onDomainMeta (const $ const "interval") (const $ const "set") "unknown" d
"set"
>>> onDomainMeta (const $ const 0) (const length) 0 d
3

Filtering

filterDomain

Arguments

:: DomainMeta
The variable's domain.

-> [RecordContent]
The original records.

-> [RecordContent]
Teh records with their variable in th domain.

Filter records based on a variable's domeain.

filterDomains

Arguments

:: [DomainMeta]
The variables' domains.

-> [RecordContent]
The original records.

-> [RecordContent]
The records with their variables in the domains.

Filter records based on variables' domains.

filterInterval

Arguments

:: VariableIdentifier
The variable's identifier.

-> (Maybe DataValue, Maybe DataValue)
The interval of values, possibly open.

-> RecordContent
The record.

-> Bool
Whether the variable in the record is in the interval.

Filter records based on an interval of values.

filterSet

Arguments

:: VariableIdentifier
The variable's identifier.

-> [DataValue]
The set of values.

-> RecordContent
The record.

-> Bool
Whether the variable in the record is in the set.

Filter records based on a set of values.