AESD.Types.Filter

Description

Types for filters on variables.

Filter trees

data Filter

A filter for records.

Constructors

Not Filter
Negation of a filter.

Union [Filter]
Union of filters.

Intersection [Filter]
Intersection of filters.

Interval VariableIdentifier (Maybe DataValue, Maybe DataValue)
Filter on a variable by a (possible open) interval of its values.

Set VariableIdentifier [DataValue]
Filter on a variable by a set of its values.

filterRecord

Arguments

:: Filter
The filter.

-> RecordContent
The record.

-> Bool
Whether the record satisfies the filter.

Apply a filter to a record.

filterRecords

Arguments

:: Filter
The filter.

-> [RecordContent]
The original records.

-> [RecordContent]
The records satisfying the filter.

Apply a filter to records.

Filter expressions

data FilterExpression

An expression for filtering. FilterExpression is used internally for encoding filters as protocol buffers.

toExpression :: Filter -> FilterExpression

Convert a filter tree to an expression for filtering.

fromExpression :: FilterExpression -> Filter

Convert an expression for filtering to a filter tree.