AESD.Types.Model

Description

Types for model metadata.

Model metadata

type ModelIdentifier = String

A unique identifier for a model.

data ModelMeta

Model metadata.

makeModelMeta

Arguments

:: ModelIdentifier
The unique identifier for the model.

-> String
The name of the model.

-> String
The URI for the model.

-> [VarMeta]
The metadata for the variables in the model.

-> [DomainMeta]
The metadata for the domains of the model's input variables, if any.

-> ModelMeta
The model metadata.

Construct model metadata.

identifier :: ModelMeta -> String

Get the model's identifier.

name :: ModelMeta -> String

Get the model's name.

uri :: ModelMeta -> String

Get the URI for the model.

varMeta :: ModelMeta -> [VarMeta]

Get the metadata for the variables in the model.

inputs :: ModelMeta -> [DomainMeta]

Get the metadata for the model's input variables, if any.

Collections of model metadata

data ModelMetas

A collection of model metadata.

makeModels :: [ModelMeta] -> ModelMetas

Make a collection of model metadata.

models :: ModelMetas -> [ModelMeta]

Get the list of model metadata from a collection.