AESD.Types.Record

Description

Types for records of data.

Values of variables

data VarValue

A value of a variable.

makeVarValue

Arguments

:: VariableIdentifier
The unique identifier for the variable.

-> DataValue
The value.

-> VarValue
The variable value.

Construct a value for a variable.

varValue :: VarValue -> (VariableIdentifier, DataValue)

Get the value for a variable.

Records

type RecordIdentifier = Int64

A unique identifier for a record.

type RecordContent = (RecordIdentifier, [(VariableIdentifier, DataValue)])

A record, consisting of a unique identifier and variables paired with their values.

data RecordData

Collections of records.

makeRecordData :: [RecordContent] -> RecordData

Construct a collection of records.

recordData :: RecordData -> [RecordContent]

Get a collection of records.

onRecordContent

Arguments

:: (Double -> a)
Handle a real value.

-> (Int64 -> a)
Handle an integer value.

-> (String -> a)
Handle a string value.

-> a
The default result value.

-> [RecordContent]
The records.

-> [(RecordIdentifier, [(VariableIdentifier, a)])]
The result of applying the handler functions to the values in the records.

Apply functions to a collection of records.

For example:

import AESD.Types.Value (integerValue, realValue, stringValue) >>> let rs = [(1, [(0, realValue 10), (1, integerValue 20), (2, stringValue "thirty")])] >>> onRecordContent show show id NA rs [(1,[(0,"10.0"),(1,"20"),(2,"thirty")])]

Filtering

filterRecords

Arguments

:: [RecordIdentifier]
The unique identifiers to be found.

-> [RecordContent]
The original records.

-> [RecordContent]
The records with the unique identifiers.

Filter records by their identifiers.

filterVariables

Arguments

:: [VariableIdentifier]
The variables to be selected.

-> [RecordContent]
The original records.

-> [RecordContent]
The selected variables from the records.

Filter records by their variables.