AESD.Types.Request

Description

Types for making requests to servers. The following requests are supported: LoadModelsMeta, LoadRecordsData, LoadBookmarkMeta, SaveBookmarkMeta, Work, and Cancel.

Requests

type RequestIdentifier = Word32

A unique identifier for a request.

data Request

A request to a server.

identifier :: Lens' Request (Maybe RequestIdentifier)

Lens for the unique identifier of a request.

subscribe :: Lens' Request Bool

Lens for whether a request is subscribing to responses.

onRequest

Arguments

:: Monad m

=> (LoadModelsMeta -> m a)
Handler loading models.

-> (LoadRecordsData -> m a)
Handler loading records.

-> (LoadBookmarkMeta -> m a)
Handler loading bookmarks.

-> (SaveBookmarkMeta -> m a)
Handler saving a bookmark.

-> (Cancel -> m a)
Handler cancelling another request.

-> (Work -> m a)
Handler performing work.

-> m a
The default result.

-> Request
The request.

-> m a
The action to handle the request.

Handle a request.

Loading models

data LoadModelsMeta

A request to load model metadata.

loadModelsMeta :: Maybe ModelIdentifier -> Request

Construct a request to load the metadata for model(s). If the model's identifier is not provided, then the request will be for all models.

onLoadModelsMeta

Arguments

:: Monad m

=> (Maybe ModelIdentifier -> m a)
Handler loading model(s).

-> LoadModelsMeta
The request.

-> m a
The action to handle the request.

Handle a request to load model(s).

Loading recordsdata

data LoadRecordsData

A request to load records.

loadRecordsData

Arguments

:: ModelIdentifier
The model's identifier.

-> Maybe Word64
The maximum number of records to return, or all records.

-> [VariableIdentifier]
The variables to select.

-> Request
The request.

Construct a request to load all records.

loadBookmarkedRecordsData

Arguments

:: ModelIdentifier
The model's identifier.

-> Maybe Word64
The maximum number of records to return, or all records.

-> [VariableIdentifier]
The variables to select.

-> BookmarkIdentifier
The unique identifier for the bookmark.

-> Request
The request.

Construct a request to load the records for a bookmark.

loadFilteredRecordsData

Arguments

:: ModelIdentifier
The model's identifier.

-> Maybe Word64
The maximum number of records to return, or all records.

-> [VariableIdentifier]
The variables to select.

-> Filter
The filter.

-> Request
The request.

Construct a request to load the records satisfying a filter.

onLoadRecordsData

Arguments

:: Monad m

=> (ModelIdentifier -> Maybe Word64 -> [VariableIdentifier] -> m a)
Handle loading all records.

-> (ModelIdentifier -> Maybe Word64 -> [VariableIdentifier] -> BookmarkIdentifier -> m a)
Handle loading bookmarked records.

-> (ModelIdentifier -> Maybe Word64 -> [VariableIdentifier] -> Filter -> m a)
Handle loading filtered recrods.

-> LoadRecordsData
The request.

-> m a
The action for handling the request.

Handle a request to load records.

onLoadRecordsData'

Arguments

:: Monad m

=> (ModelIdentifier -> Maybe Word64 -> [VariableIdentifier] -> Maybe (Either BookmarkIdentifier Filter) -> m a)
Handle loading records.

-> LoadRecordsData
The request.

-> m a
Action to handle the request.

Handle a request to load records.

Loading bookmarks

data LoadBookmarkMeta

A request to load bookmark metadata.

loadBookmarkMeta

Arguments

:: ModelIdentifier
The model's identifier.

-> Maybe BookmarkIdentifier
The bookmark, or all bookmarks.

-> Request
The request.

Construct a request to load bookmark metadata.

onLoadBookmarkMeta

Arguments

:: Monad m

=> (ModelIdentifier -> Maybe BookmarkIdentifier -> m a)
Handle loading bookmark(s).

-> LoadBookmarkMeta
The request.

-> m a
The action to handle the request.

Handle a request to load bookmark(s).

Saving bookmarks

data SaveBookmarkMeta

A request to save a bookmark.

saveBookmarkMeta

Arguments

:: ModelIdentifier
The model's identifier.

-> BookmarkMeta
The bookmark's metadata.

-> Request
The request.

Construct a request to save a bookmark.

onSaveBookmarkMeta

Arguments

:: Monad m

=> (ModelIdentifier -> BookmarkMeta -> m a)
Handle saving a bookmark.

-> SaveBookmarkMeta
The request.

-> m a
The action for handling the request.

Handle a request to save a bookmark.

Creating work

data Work

A request for new work.

work

Arguments

:: ModelIdentifier
The model's identifier.

-> [VarValue]
The values of input variables.

-> Request
The request.

Construct a request for new work.

onWork

Arguments

:: Monad m

=> (ModelIdentifier -> [VarValue] -> m a)
Handle doing new work.

-> Work
The request.

-> m a
The action to handle the request.

Handle a request for new work.

Canceling requests

data Cancel

A request to cancel a previous request.

cancel

Arguments

:: RequestIdentifier
The identifier for the request to be cancelled.

-> Request
The request.

Construct a cancelation request.

onCancel

Arguments

:: Monad m

=> (RequestIdentifier -> m a)
Handle cancelling.

-> Cancel
The request.

-> m a
The action to handle the request.

Handle a request for cancellation.