AESD.Types.Value

Description

Types for data values.

Data values

data DataValue

A data value.

realValue :: Double -> DataValue

Construct a real value.

integerValue :: Int64 -> DataValue

Construct an integer value.

stringValue :: String -> DataValue

Construct a string value.

onDataValue

Arguments

:: (Double -> a)
Handle a real value.

-> (Int64 -> a)
Handle an integer value.

-> (String -> a)
Handle a string value.

-> a
The default result.

-> DataValue
The data value.

-> a
The result of applying the first applicable function to the value.

Apply a function to a value.

>>> import AESD.Types.Value (onDataValue, realValue)
>>> 
>>> let x = realValue 42
>>> onDataValue (("real value = " ++) . show) (("integer value = " ++) . show) ("string value = " ++ ) "unknown value" x
"real value = 42.0"

onDataValue2

Arguments

:: (Double -> Double -> a)
Handle a pair of real values.

-> (Int64 -> Int64 -> a)
Handle a pair of integer values.

-> (String -> String -> a)
Handle a pair of string values.

-> a
The default result.

-> DataValue
The first value.

-> DataValue
The resecond value.

-> a
The result of applying the first applicable function to the pair of values.

Apply a binary function to a pair of values.

onDataValues

Arguments

:: ([Double] -> a)
Handle a list of real values.

-> ([Int64] -> a)
Handle a list of integer values.

-> ([String] -> a)
Handle a list of string values.

-> a
The default result.

-> [DataValue]
The data values.

-> a

Apply a function to a list of values.

Variable types

data VarType

Types of variables.

Constructors

RealVar
Type for a real value.

IntegerVar
Type for an integer value.

StringVar
Type for a string value.

varType :: DataValue -> Maybe VarType

Get the type of a value.

varTypes :: DataValue -> [VarType]

Get the possible types of a value.

sameVarTypes :: [DataValue] -> Bool

Determine whether a list of values are of the same type.

commonVarType :: [DataValue] -> VarType

Determine the commonly expressible type of a list of values.

commonVarTypes :: [[DataValue]] -> [VarType]

Determine the list of commonly expressible types of a table of values.

castVarType :: VarType -> DataValue -> DataValue

Cast a value to a particular type.

castVarTypes :: [DataValue] -> (VarType, [DataValue])

Cast a list of values to a commmonly expressible type.