Network.WebSockets.STM

Description

WebSockets with STM queues for sending and receiving messages. There is one thread for sending messages, another thread for receiveing them, and separate threads to process the messages received from each request. Each worker can also have its own thread.

Types

data Communicator s r k

The state of communication via WebSockets.

Control

start

Arguments

:: (WebSocketsData s, WebSocketsData r, Ord k)

=> Connection
The WebSockets connection.

-> (k -> s -> s)
Function for adding the key to a message.

-> (r -> k)
Function for getting the key of a message.

-> IO (Communicator s r k)
Action to create the threads and start communication.

Fork the threads needed for communication.

stop :: Communicator s r k -> IO ()

Kill the communication threads.

waitToStop :: Communicator s r k -> IO ()

Wait for the WebSockets connection to be closed.

complete

Arguments

:: Ord k

=> Communicator s r k
The state of communication.

-> k
The message key.

-> IO ()
Action for stopping the processing of messages with the key.

Stop processing messages with a particular key.

Communication

send

Arguments

:: Communicator s r k
The state of communication.

-> k
The message key.

-> s
The message to be sent.

-> IO ()
Action for enqueuing the message for sending.

Send a message.

launch

Arguments

:: Ord k

=> Communicator s r k
The state of communication.

-> k
The message key.

-> (r -> IO Bool)
Handler for messages, which returns whether the handling of this message key is complete.

-> IO ()
Action for processing incoming messages.

Process messages for a given key.

Computation

forkWork :: Communicator s r k -> IO () -> IO ()

Fork a worker thread.